+++ AdventOfCode/2019/Year2019Day5.cs
@@ -7,7 +7,7 @@ public class Year2019Day5 : IDay
 {
 	public async Task<string> RunSolution1Async(IList<string> input)
 	{
-		var code = input.First().Split(',').Select(int.Parse).ToList();
+		var code = input.First().Split(',').Select(long.Parse).ToList();
 		Intcomputer intComputer = new(code);
 		await intComputer.RunAsync();
 		Console.WriteLine(string.Join("",intComputer.GetOutputBuffer()));
@@ -17,7 +17,7 @@ public class Year2019Day5 : IDay
 
 	public async Task<string> RunSolution2Async(IList<string> input)
 	{
-		var code = input.First().Split(',').Select(int.Parse).ToList();
+		var code = input.First().Split(',').Select(long.Parse).ToList();
 		Intcomputer intComputer = new(code);
 		await intComputer.RunAsync();
 		Console.WriteLine(string.Join("",intComputer.GetOutputBuffer()));
